修复报错推送报错

+++ AiCard/AiCard/Bll/VIPAccountBLL.cs
@@ -120,20 +120,29 @@ namespace AiCard.Bll
                                 result.retCode = ReqResultCode.success;
                                 result.retMsg = "计算佣金成功";
                             }
-                            var vUser = db.Users.FirstOrDefault(s => s.Id == vUserID);
-                            switch (logType)
+                            try
                             {
-                                case Common.Enums.VipAmountLogType.NewCard:
-                                    WeChatNofity(parentUser.UserID, parentProfitAmount, $"{vUser.NickName}注册了名片得到了{parentProfitAmount}元奖励");
-                                    break;
-                                case Common.Enums.VipAmountLogType.NewChild2nd:
-                                    WeChatNofity(parentUser.UserID, parentProfitAmount, $"{vUser.NickName}成为了您的一级会员得到了{parentProfitAmount}元奖励");
-                                    break;
+                                var vUser = db.Users.FirstOrDefault(s => s.Id == vUserID);
+                                switch (logType)
+                                {
+                                    case Common.Enums.VipAmountLogType.NewCard:
+                                        WeChatNofity(parentUser.UserID, parentProfitAmount, $"{vUser.NickName}注册了名片得到了{parentProfitAmount}元奖励");
+                                        break;
+                                    case Common.Enums.VipAmountLogType.NewChild2nd:
+                                        WeChatNofity(parentUser.UserID, parentProfitAmount, $"{vUser.NickName}成为了您的一级会员得到了{parentProfitAmount}元奖励");
+                                        break;
+                                }
+                                
+                                if (grandfatherUser != null)
+                                {
+                                    WeChatNofity(grandfatherUser.UserID, GrandfatheredProfitAmount, $"{vUser.NickName}成为了您的二级会员得到了{GrandfatheredProfitAmount}元奖励");
+                                }
                             }
-                            if (grandfatherUser != null)
+                            catch (Exception)
                             {
-                                WeChatNofity(grandfatherUser.UserID, GrandfatheredProfitAmount, $"{vUser.NickName}成为了您的二级会员得到了{GrandfatheredProfitAmount}元奖励");
+
                             }
+
                         }
                         else
                         {
@@ -172,10 +181,14 @@ namespace AiCard.Bll
                 var form = db.WeChatMiniNotifyForms.FirstOrDefault(s => s.AppID == config.AppID
                     && s.UserID == userID
                     && s.EndDateTime > DateTime.Now);
-                var temp = new Common.WeChat.WeChatMessageTemp.PReceivableNotifyWeChatMessage(amount, content, DateTime.Now);
-                wechat.SendMessage(openID, form.FormID, null, temp);
-                db.WeChatMiniNotifyForms.Remove(form);
-                db.SaveChanges();
+                if (form != null)
+                {
+                    var temp = new Common.WeChat.WeChatMessageTemp.PReceivableNotifyWeChatMessage(amount, content, DateTime.Now);
+                    wechat.SendMessage(openID, form.FormID, null, temp);
+                    db.WeChatMiniNotifyForms.Remove(form);
+                    db.SaveChanges();
+                }
+              
             }
 
         }
